Split protocols and classes in separate files
