+++ Sources/Enums/KeyboardStatus.swift
@@ -1,0 +1,13 @@
+//
+//  KeyboardStatus.swift
+//  KeyboardHandler-iOS
+//
+//  Created by Andrei Nastasiu on 05/02/2018.
+//  Copyright © 2018 Dolfn. All rights reserved.
+//
+
+import Foundation
+
+public enum KeyboardStatus {
+    case showing, hiding
+}
