+++ Sources/Protocols/KeyboardFirstResponder.swift
@@ -1,0 +1,13 @@
+//
+//  KeyboardFirstResponder.swift
+//  KeyboardHandler-iOS
+//
+//  Created by Andrei Nastasiu on 05/02/2018.
+//  Copyright © 2018 Dolfn. All rights reserved.
+//
+
+import Foundation
+
+protocol KeyboardFirstResponder {
+    var isTextInputFirstResponder: Bool { get }
+}
