+++ Sources/Protocols/KeyboardHandlerDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  KeyboardHandlerDelegate.swift
+//  KeyboardHandler-iOS
+//
+//  Created by Andrei Nastasiu on 05/02/2018.
+//  Copyright © 2018 Dolfn. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+protocol KeyboardHandlerDelegate: class {
+    func didShowKeyboard(_ keyboardHeight: CGFloat)
+}
