+++ Sources/Protocols/TapGestureRecognizerManagerDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  TapGestureRecognizerManagerDelegate.swift
+//  KeyboardHandler-iOS
+//
+//  Created by Andrei Nastasiu on 05/02/2018.
+//  Copyright © 2018 Dolfn. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+protocol TapGestureRecognizerManagerDelegate: class {
+    func gestureRecognizerManagerDidTapOnView(_ tapGestureRecognizerManager: TapGestureRecognizerManager)
+}
